requirements.txt
torch
torchvision
numpy
pyyaml
tqdm
pillow
scikit-learn
